# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Nia API
  version: 2.0.0
  summary: Control-plane API for Nia's realtime language tutor
  description: |
    Nia's API authenticates learners, supplies initial tutor settings, issues
    short-lived OpenAI Realtime client secrets, stores text turns, and creates
    post-session feedback. Audio travels directly between the client and OpenAI
    over WebRTC; it is not uploaded to this API. A connected direct client can
    update its provider session, so initial tutor settings are not an immutable
    authorization boundary.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - FirebaseBearer: []
    FirebaseAppCheck: []
  - LocalDemoBearer: []
paths:
...
  /api/v1/me/preferences:
    get:
      tags: [Profile]
      operationId: getPreferences
      summary: Get learner preferences
      responses:
        '200':
          description: Current preferences, including server defaults when none were saved.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preferences'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Profile]
      operationId: updatePreferences
      summary: Update learner preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferencesPatch'
      responses:
        '200':
          description: Updated preferences.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preferences'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/realtime/sessions:
    post:
      tags: [Realtime]
      operationId: createRealtimeSession
      summary: Start a realtime tutoring session
      description: |
        Creates a conversation and returns connection metadata. Production mode
        includes a short-lived OpenAI client secret. Credential-free demo mode
        returns `client_secret: null` and `transport: demo`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRealtimeSessionRequest'
            example:
              preferences:
                target_language: es
                level: intermediate
                topic: Ordering dinner at a restaurant
                correction_style: summary
      responses:
        '201':
          description: Realtime session created.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRealtimeSessionResponse'
              examples:
                localDemo:
                  summary: Credential-free local session
                  value:
                    conversation:
                      id: conv_01JNA6HF7BP6KS7X1Q43KM9N33
                      status: active
                      preferences:
                        target_language: es
                        level: intermediate
                        topic: Ordering dinner at a restaurant
                        correction_style: summary
                      turn_count: 0
                      created_at: '2026-07-16T09:00:00Z'
                      updated_at: '2026-07-16T09:00:00Z'
                    client_secret: null
                    realtime:
                      transport: demo
                      endpoint: demo://local
                      model: deterministic-demo
                productionWebRTC:
                  summary: Production WebRTC bootstrap
                  value:
                    conversation:
                      id: conv_01JNA6HF7BP6KS7X1Q43KM9N34
                      status: active
                      preferences:
                        target_language: es
                        level: intermediate
                        topic: Ordering dinner at a restaurant
                        correction_style: summary
                      turn_count: 0
                      created_at: '2026-07-16T09:00:00Z'
                      updated_at: '2026-07-16T09:00:00Z'
                    client_secret:
                      value: ephemeral-example-value
                      expires_at: 1784193000
                    realtime:
                      transport: webrtc
                      endpoint: https://api.openai.com/v1/realtime/calls
                      model: gpt-realtime-2.1
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/conversations:
    get:
      tags: [Conversations]
      operationId: listConversations
      summary: List the learner's conversations
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Conversations ordered from most recently updated.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/conversations/{conversation_id}:
    parameters:
      - $ref: '#/components/parameters/ConversationID'
    get:
      tags: [Conversations]
      operationId: getConversation
      summary: Get a conversation and its feedback
      responses:
        '200':
          description: Conversation detail.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDetail'
              examples:
                demoFeedback:
                  summary: Completed local demo conversation
                  value:
                    conversation:
                      id: conv_01JNA6HF7BP6KS7X1Q43KM9N33
                      status: completed
                      preferences:
                        target_language: es
                        level: intermediate
                        topic: Ordering dinner at a restaurant
                        correction_style: summary
                      turn_count: 2
                      created_at: '2026-07-16T09:00:00Z'
                      updated_at: '2026-07-16T09:05:00Z'
                    turns:
                      - id: turn_demo_0001
                        role: user
                        text: Yo soy bien y quiero una mesa para dos.
                        occurred_at: '2026-07-16T09:00:00Z'
                      - id: turn_demo_0002
                        role: assistant
                        text: 'Una corrección: en este caso, di “Estoy bien”.'
                        occurred_at: '2026-07-16T09:00:02Z'
                    feedback:
                      summary: You completed 1 Spanish learner turn about Ordering dinner at a restaurant.
                      strengths:
                        - You kept the exchange moving for 1 learner turn.
                        - 'Your latest idea was specific: “Yo soy bien.”'
                      corrections:
                        - original: Yo soy bien.
                          corrected: Estoy bien.
                          explanation: Use “estar” for a temporary state or feeling.
                      next_steps:
                        - Responde otra vez añadiendo cuándo, dónde y por qué.
                        - Practica una pregunta de seguimiento sobre ordering dinner at a restaurant.
                      generated_at: '2026-07-16T09:05:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Conversations]
      operationId: deleteConversation
      summary: Delete a conversation
      description: |
        Makes the conversation unavailable immediately, then removes its live
        transcript and generated feedback. Cloud backups, point-in-time recovery,
        and external processors remain subject to the operator's retention policy.
      responses:
        '204':
          description: Conversation deleted.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/conversations/{conversation_id}/turns/{turn_id}:
    parameters:
      - $ref: '#/components/parameters/ConversationID'
      - $ref: '#/components/parameters/TurnID'
    put:
      tags: [Conversations]
      operationId: upsertConversationTurn
      summary: Store a transcript turn idempotently
      description: |
        Reusing a turn ID replaces that turn instead of duplicating it. New
        turns are accepted for two hours after conversation creation and each
        conversation is capped atomically at 200 turns. Existing IDs remain
        retryable at the cap.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertTurnRequest'
      responses:
        '200':
          description: Turn stored.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Turn'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/conversations/{conversation_id}/complete:
    parameters:
      - $ref: '#/components/parameters/ConversationID'
    post:
      tags: [Conversations]
      operationId: completeConversation
      summary: Complete a conversation and generate feedback
      description: |
        Requires at least one stored learner/user turn. An assistant greeting by
        itself cannot be completed or consume feedback-generation quota. The
        operation is idempotent after feedback has been persisted successfully.
        While feedback generation or deletion owns the conversation lease, the
        API returns a retryable conflict.
      responses:
        '200':
          description: Completed conversation with feedback.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    FirebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
      description: Firebase ID token required by production protected routes.
    FirebaseAppCheck:
      type: apiKey
      in: header
      name: X-Firebase-AppCheck
      description: Valid Firebase App Check token required together with the production bearer token.
    LocalDemoBearer:
      type: http
      scheme: bearer
      bearerFormat: local demo token
      description: Fixed local demo identity. Production mode always rejects this credential.
  headers:
    RequestID:
      description: Correlation ID generated or accepted by the API.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        example: 01JNA6HF7BP6KS7X1Q43KM9N33
  parameters:
    ConversationID:
      name: conversation_id
      in: path
      required: true
      description: Server-issued conversation identifier.
      schema:
        $ref: '#/components/schemas/Identifier'
    TurnID:
      name: turn_id
      in: path
      required: true
      description: Client-issued idempotency identifier for a transcript turn.
      schema:
        $ref: '#/components/schemas/Identifier'
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned by the previous page.
      schema:
        type: string
        minLength: 1
        maxLength: 512
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of conversations to return.
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
  responses:
    BadRequest:
      description: The request is malformed or failed validation.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Authentication is missing or invalid.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: The resource does not exist or is not owned by this learner.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: The resource state does not allow this operation.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    RateLimited:
      description: The learner exceeded a configured request-rate limit.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalError:
      description: An unexpected server error occurred.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    BadGateway:
      description: A required provider returned an invalid or rejected response.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unavailable:
      description: A required dependency is temporarily unavailable.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    Preferences:
      type: object
      additionalProperties: false
      required: [target_language, level, topic, correction_style]
      properties:
        target_language:
          type: string
          description: Language practised in the current product version.
          enum: [en, es, ca]
          example: es
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        topic:
          type: string
          minLength: 1
          maxLength: 120
          example: Ordering dinner at a restaurant
        correction_style:
          type: string
          enum: [gentle, immediate, summary]
    PreferencesPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        target_language:
          type: string
          enum: [en, es, ca]
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        topic:
          type: string
          minLength: 1
          maxLength: 120
        correction_style:
          type: string
          enum: [gentle, immediate, summary]
    CreateRealtimeSessionRequest:
      type: object
      additionalProperties: false
      required: [preferences]
      properties:
        preferences:
          $ref: '#/components/schemas/Preferences'
    CreateRealtimeSessionResponse:
      type: object
      additionalProperties: false
      required: [conversation, client_secret, realtime]
      properties:
        conversation:
          $ref: '#/components/schemas/ConversationSummary'
        client_secret:
          oneOf:
            - $ref: '#/components/schemas/ClientSecret'
            - type: 'null'
        realtime:
          $ref: '#/components/schemas/RealtimeConnection'
...
    ConversationList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ConversationSummary'
        next_cursor:
          type: string
          minLength: 1
          maxLength: 512
          description: Omitted when there is no next page.
    ConversationDetail:
      type: object
      additionalProperties: false
      required: [conversation, turns, feedback]
      properties:
        conversation:
          $ref: '#/components/schemas/ConversationSummary'
        turns:
          type: array
          items:
            $ref: '#/components/schemas/Turn'
        feedback:
          oneOf:
            - $ref: '#/components/schemas/Feedback'
            - type: 'null'
...